requirements.txt
altair==5.3.0
asgiref==3.8.1
asttokens==2.4.1
attrs==23.2.0
blinker==1.8.2
boto3==1.35.99
botocore==1.35.99
cachetools==5.3.3
certifi==2024.7.4
channels==4.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.2
contourpy==1.2.1
cv==1.0.0
cycler==0.12.1
debugpy==1.8.2
decorator==5.1.1
distlib==0.3.8
dj-database-url==2.3.0
Django==5.0.7
django-cors-headers==4.6.0
djangorestframework==3.15.2
executing==2.0.1
filelock==3.15.4
fonttools==4.51.0
gitdb==4.0.11
GitPython==3.1.43
gunicorn==23.0.0
idna==3.7
imageio==2.34.0
ipykernel==6.29.5
ipython==8.26.0
jedi==0.19.1
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
kiwisolver==1.4.5
lazy_loader==0.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.4
matplotlib-inline==0.1.7
mdurl==0.1.2
nest-asyncio==1.6.0
networkx==3.3
nltk==3.8.1
numpy==1.26.4
opencv-python==4.9.0.80
packaging==24.0
pandas==2.2.2
parso==0.8.4
pillow==10.3.0
platformdirs==4.2.2
prompt_toolkit==3.0.47
protobuf==5.27.2
psutil==6.0.0
psycopg2-binary==2.9.10
pure-eval==0.2.2
pyarrow==16.1.0
pydeck==0.9.1
Pygments==2.18.0
pyparsing==3.1.2
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.1
pywin32==306
pyzmq==26.0.3
referencing==0.35.1
regex==2024.5.15
requests==2.32.3
rich==13.7.1
rpds-py==0.19.0
s3transfer==0.10.4
scikit-image==0.23.1
scikit-learn==1.5.1
scipy==1.13.0
six==1.16.0
smmap==5.0.1
sqlparse==0.5.1
stack-data==0.6.3
streamlit==1.36.0
tenacity==8.5.0
tesseract==0.1.3
threadpoolctl==3.5.0
tifffile==2024.2.12
toml==0.10.2
toolz==0.12.1
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
virtualenv==20.26.3
waitress==3.0.2
watchdog==4.0.1
wcwidth==0.2.13
whitenoise==6.8.2
